% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{add_all_row}
\alias{add_all_row}
\title{Add a total "All" row}
\usage{
add_all_row(tbl)
}
\arguments{
\item{tbl}{The output of \code{summarise_table1()}.}
}
\value{
A tibble identical to the input but with one additional
row summarising all infection types.
}
\description{
Appends a summary row where all infection types are combined
into a single total, with \code{infection_type = "All"}.
}
