% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{choose_sample_cases}
\alias{choose_sample_cases}
\title{Choose which dataset to analyse (Germany vs EU/EEA)}
\usage{
choose_sample_cases(sample = c("Germany", "EU/EEA"))
}
\arguments{
\item{sample}{Either \code{"Germany"} or \code{"EU/EEA"}.}
}
\value{
A data frame of simulated cases.
}
\description{
Convenience helper for the Shiny app. Returns either the German
simulated dataset (\code{shayhai_cases}) or the EU/EEA simulated
dataset (\code{shayhai_cases_ecdc}), depending on user input.
}
\examples{
# Get the German PPS-like dataset
choose_sample_cases("Germany")

# Get the EU/EEA PPS-like dataset
choose_sample_cases("EU/EEA")

# If you don't supply an argument, it will default to "Germany":
choose_sample_cases()

}
